+++ CHANGELOG.md
@@ -3981,3 +3981,4 @@
 - [2026-08-28 10:47] docs(security): update readme and API contracts for security
 - [2026-08-28 11:22] test(analytics): add unit tests for analytics business logic
 - [2026-08-28 11:26] feat(test): add bi-directional event bus handler for test
+- [2026-08-28 11:48] docs(theme): update readme and API contracts for theme